[Model] User and hub models
Add user and hub models to the project. User model is the standard
Devise-generated user without any of the fancy additional features. The
hub model comes with BDD-style unit and functional tests. Request tests
are being ignored for now.

- Adds mandatory authentication to the hub controller.

- Adds FactoryGirl with respective factories for both models.

The scaffold generates a ton of view-specific files which we need to
decide what to do about. Might not need them at all, in which case they
should be deleted right away.