Convert SensorReading to accept mac address
The SensorReading POST route now takes a sensor_id, a mac_address and
a watthours value. The former two are used to uniquely pin down a
sensor for which the reading is being reported.

Note: doesn't contain any sort of elaborate error checking and input
validation. Will be added later per-need.