Add seeding logic for a watthour receiver account
Adds seeding logic for a sample account, should be ideal for testing
against Heroku.

Additionally fixes issue with sensor id being primary key. There is now
an additional column called "local_id" on sensors table and
"sensor_local_id" on the sensor_readings table that is used to link
the two tables. The issue is that Rails doesn't support composite
primary keys by default, or hub_id a per-hub id, so this is the
workaround until we start coding directly against a DB system and
can use it natively. I would have preferred hub_id + id, but it's not
worth the enormous hassle just yet.