Add Spork support
Adds Spork to the project and integrates it with Guard. Assuming the
bundle exec fix was applied, running "guard" in the terminal will now
launch an instance of spork and have guard test against that instance.

Additionally turns off NOT re-running everything when a failing test
passes. At this point there are not enough tests to justify that
optimization and it catches a lot of pesky issues this way. We will
likely turn that off if the project gets huge.