Adds CanCan abilities support into the project
CanCan gem is used to give different permissions to project sager
users. At this point the roles table is not populated with any
roles, and the default role is given the expected restricted
permissions:

- Manage new and own hubs

- Manage new and own sensors

- Read one's own sensor readings

Adds model-level tests to ensure that the gem permissions are set as
intended. Might want to consider controller gems, although rbates
recommends to keep it very simple.