Adds delayed_job and sensor_reading generation
Delayed job is used to schedule background tasks, in our case the
sensor reading generation job. Foreman is used to spawn background
Rails processes that execute those tasks.

As of right now, a user with the 'demo' column set as true, must
have exactly 4 sensors. If all of this is true, then the generator
will, once a minute, add a reading to each one of the sensors. The
sensor order is hardcoded at this point (computer, microwave.. and so
on).