Switches db from sqlite3 to postgresql
The app now REQUIRES postgresql to work properly. You need to setup
a postgresql server on the development box and make sure it is running
and accessible at the address:port specified in the database.yml file.

It is also imperative that the server has a rails_dev user created
in it with the permission to create new databases. This article
was the best at guiding through the process of postgresql config:

http://mrfrosti.com/2011/06/setup-postgresql-with-rails-on-linux/

Note that the user needs to be created with a password (use the -P)
switch, unlike in the tutorial. The password is specified in the
same database.yml file.