Workaround for client authorization
The client currently has no way of authenticating itself with the web
app and thus needs to be able to get all of the sensors in the system
without any additional work. This enables the scenario, with the
downside of letting anybody on the net have read-only access to
the full list of sensors.